+++ packages/@tinacms/styles/src/FontLoader.tsx
@@ -1,6 +1,6 @@
 /**
 
-Copyright 2019 Forestry.io Inc
+Copyright 2021 Forestry.io Holdings, Inc.
 
 
 Licensed under the Apache License, Version 2.0 (the "License");