+++ packages/react-tinacms-inline/src/blocks/add-block-menu.tsx
@@ -1,6 +1,6 @@
 /**
 
-Copyright 2019 Forestry.io Inc
+Copyright 2021 Forestry.io Holdings, Inc.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -21,6 +21,7 @@ import styled, { css } from 'styled-components'
 import { BlockTemplate } from 'tinacms'
 import { IconButton } from '@tinacms/styles'
 import { AddIcon } from '@tinacms/icons'
+import { Input } from 'tinacms'
 
 import { getOffset, getOffsetX, getOffsetY } from '../styles'
 
@@ -41,7 +42,9 @@ export function AddBlockMenu({
 }: AddBlockMenuProps) {
   const [isOpen, setIsOpen] = React.useState(false)
   const addBlockButtonRef = React.useRef<HTMLButtonElement>(null)
+  const addBlockMenuRef = React.useRef<HTMLDivElement>(null)
   const [openTop, setOpenTop] = React.useState(false)
+  const [filterValue, setFilterValue] = React.useState('')
 
   const getDefaultProps = (defaultItem: any) => {
     return typeof defaultItem === 'function' ? defaultItem() : defaultItem
@@ -51,6 +54,7 @@ export function AddBlockMenu({
     event.stopPropagation()
     event.preventDefault()
     const addBlockButtonElem = addBlockButtonRef.current
+    setFilterValue('')
 
     if (addBlockButtonElem !== null) {
       const menuBounding = addBlockButtonElem.getBoundingClientRect()
@@ -77,10 +81,17 @@ export function AddBlockMenu({
   }
 
   React.useEffect(() => {
-    const inactivateBlockMenu = () => setIsOpen(false)
-    document.addEventListener('mouseup', inactivateBlockMenu, false)
-    return () => document.removeEventListener('mouseup', inactivateBlockMenu)
-  }, [])
+    const inactivateBlockMenu = (event: any) => {
+      if (
+        addBlockMenuRef.current &&
+        !addBlockMenuRef.current.contains(event.target)
+      ) {
+        setIsOpen(false)
+      }
+    }
+    document.addEventListener('mousedown', inactivateBlockMenu)
+    return () => document.removeEventListener('mousedown', inactivateBlockMenu)
+  }, [addBlockButtonRef])
 
   return (
     <AddBlockWrapper
@@ -98,32 +109,59 @@ export function AddBlockMenu({
       >
         <AddIcon />
       </AddBlockButton>
-      <BlocksMenu openTop={openTop} isOpen={isOpen}>
-        {Object.keys(blocks).map((key: string) => {
-          const template = blocks[key].template
+      <BlocksMenu openTop={openTop} isOpen={isOpen} ref={addBlockMenuRef}>
+        {Object.keys(blocks).length > 9 && (
+          <DropdownHeader>
+            <SelectFilter
+              placeholder="Filter"
+              onChange={event => setFilterValue(event.target.value)}
+              onClick={(event: any) => {
+                event.preventDefault()
+                event.stopPropagation()
+              }}
+              value={filterValue}
+            />
+          </DropdownHeader>
+        )}
+        <BlocksMenuOptions>
+          {Object.keys(blocks).filter(key => {
+            const label = blocks[key].template.label
+            return label.toLowerCase().includes(filterValue.toLowerCase())
+          }).length > 0 ? (
+            Object.keys(blocks)
+              .filter(key => {
+                const label = blocks[key].template.label
+                return label.toLowerCase().includes(filterValue.toLowerCase())
+              })
+              .map((key: string) => {
+                const template = blocks[key].template
 
-          if (!template) {
-            console.error(`No template for ${key} block exists`)
+                if (!template) {
+                  console.error(`No template for ${key} block exists`)
 
-            return null
-          } else {
-            return (
-              <BlockOption
-                key={template?.label}
-                onClick={event => {
-                  event.stopPropagation()
-                  event.preventDefault()
-                  addBlock({
-                    _template: key,
-                    ...getDefaultProps(template?.defaultItem),
-                  })
-                }}
-              >
-                {template?.label}
-              </BlockOption>
-            )
-          }
-        })}
+                  return null
+                } else {
+                  return (
+                    <BlockOption
+                      key={template?.label}
+                      onClick={event => {
+                        event.stopPropagation()
+                        event.preventDefault()
+                        addBlock({
+                          _template: key,
+                          ...getDefaultProps(template?.defaultItem),
+                        })
+                      }}
+                    >
+                      {template?.label}
+                    </BlockOption>
+                  )
+                }
+              })
+          ) : (
+            <BlockOption disabled>No blocks to display</BlockOption>
+          )}
+        </BlocksMenuOptions>
       </BlocksMenu>
     </AddBlockWrapper>
   )
@@ -216,8 +254,9 @@ const AddBlockWrapper = styled.div<AddBlockWrapperProps>(p => {
 
 const BlocksMenu = styled.div<AddMenuProps>`
   min-width: 192px;
-  border-radius: var(--tina-radius-big);
+  border-radius: var(--tina-radius-small);
   border: 1px solid var(--tina-color-grey-2);
+  border-bottom-color: var(--tina-color-grey-3);
   display: flex;
   flex-direction: column;
   align-items: stretch;
@@ -257,6 +296,41 @@ const BlocksMenu = styled.div<AddMenuProps>`
     `};
 `
 
+const BlocksMenuOptions = styled.div<AddMenuProps>`
+  min-width: 192px;
+  max-height: 25vh;
+  display: flex;
+  flex-direction: column;
+  align-items: stretch;
+  justify-content: flex-start;
+  background-color: white;
+  overflow: auto;
+  background:
+		/* Shadow covers */ linear-gradient(
+      white 30%,
+      rgba(255, 255, 255, 0)
+    ),
+    linear-gradient(rgba(255, 255, 255, 0), white 70%) 0 100%,
+    /* Shadows */
+      linear-gradient(
+        to bottom,
+        var(--tina-color-grey-2),
+        var(--tina-color-grey-1),
+        rgba(0, 0, 0, 0)
+      ),
+    linear-gradient(
+        to top,
+        var(--tina-color-grey-3),
+        var(--tina-color-grey-1),
+        rgba(0, 0, 0, 0)
+      )
+      0 100%;
+  background-repeat: no-repeat;
+  background-color: white;
+  background-size: 100% 84px, 100% 84px, 100% 28px, 100% 28px;
+  background-attachment: local, local, scroll, scroll;
+`
+
 const BlockOption = styled.button`
   all: unset;
   box-sizing: border-box;
@@ -267,6 +341,7 @@ const BlockOption = styled.button`
   font-size: var(--tina-font-size-1);
   padding: 0 12px;
   height: 34px;
+  flex: 0 0 auto;
   font-weight: var(--tina-font-weight-regular);
   width: 100%;
   background: none;
@@ -278,9 +353,39 @@ const BlockOption = styled.button`
 
   &:hover {
     color: var(--tina-color-primary);
-    background-color: #f6f6f9;
+    background-color: rgba(0, 0, 0, 0.03);
   }
   &:not(:last-child) {
-    border-bottom: 1px solid #efefef;
+    border-bottom: 1px solid rgba(0, 0, 0, 0.05);
+  }
+
+  ${props =>
+    props.disabled &&
+    css`
+      pointer-events: none;
+      color: var(--tina-color-grey-5);
+      background: var(--tina-color-grey-1);
+
+      &:hover {
+        color: var(--tina-color-grey-8);
+        background-color: var(--tina-color-grey-1);
+      }
+    `};
+`
+
+const DropdownHeader = styled.div`
+  position: sticky;
+  top: 0;
+  padding: 10px;
+  background-color: var(--tina-color-grey-1);
+  border-bottom: 1px solid var(--tina-color-grey-2);
+`
+
+const SelectFilter = styled(Input)`
+  height: 36px;
+  flex: 0 1 auto;
+
+  ::placeholder {
+    color: var(--tina-color-grey-4);
   }
 `