+++ packages/react-tinacms-inline/src/index.ts
@@ -1,6 +1,6 @@
 /**
 
-Copyright 2019 Forestry.io Inc
+Copyright 2021 Forestry.io Holdings, Inc.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -33,3 +33,4 @@ export * from './inline-group'
 export { InlineSettings } from './inline-settings'
 export * from './blocks'
 export * from './styles'
+export * from './rbie'