+++ packages/tinacms/src/components/media/pagination.tsx
@@ -1,6 +1,6 @@
 /**
 
-Copyright 2019 Forestry.io Inc
+Copyright 2021 Forestry.io Holdings, Inc.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -48,8 +48,13 @@ export function PageLinks({ list, setOffset }: MediaPaginatorProps) {
 
   for (let i = 1; i <= numPages; i++) {
     const active = i === currentPageIndex
+    const nextOffset = (i - 1) * limit
     pageLinks.push(
-      <PageNumber active={active} onClick={() => setOffset(i * limit)}>
+      <PageNumber
+        key={`page-${i}`}
+        active={active}
+        onClick={() => setOffset(nextOffset)}
+      >
         {i}
       </PageNumber>
     )